legacy_archive/requirements.txt
Flask==3.0.0
pandas==2.1.4
numpy==1.26.3
xgboost==2.0.3
scikit-learn==1.4.0
python-dotenv==1.0.0
colorlog==6.8.0
